requirements.txt
google-generativeai==0.3.2
pandas==2.1.4
python-dotenv==1.0.0
streamlit==1.28.1
tabulate==0.9.0
